requirements.txt
azure-identity==1.10.0
msal==1.22.0
python-dotenv==1.0.1
azure-cli==2.50.0
msal-extensions==1.0.0
jinja2==3.1.4
pyjwt==2.8.0